+++ include/TileMap.h
@@ -17,7 +17,7 @@ public:
 private:
 	std::vector<std::vector<MetaTile>> m_tileMap;
 	// real background accurate tilemap based on sub tiles made by grit
-	static u8 m_bgTileMap[SUB_TILE::COUNT_W][SUB_TILE::COUNT_H];
+	static u16 m_bgTileMap[SUB_TILE::COUNT_W][SUB_TILE::COUNT_H];
 
 	//Convert m_tileMap into a background adapted tilemap
 	void convertMap();
